Move delete_index function into by_collection.py
